Split user profile page component to smaller components
