+++ components/user-profile/user-skills.tsx
@@ -1,113 +1,10 @@
-import { UserProfile } from "lib/airtable/user-profile";
-import { Layout, Section, SectionContent } from "components/layout";
-import { Body, Heading1, Heading2 } from "./typography";
-import { Button } from "./buttons";
+import { Body, Heading2 } from "components/typography";
 import { Skill } from "lib/airtable/skills";
 import { unique } from "lib/utils";
 import { useState } from "react";
 import styled from "styled-components";
 
-export type UserProfilePageState =
-  | "loading"
-  | "signed_out"
-  | "loading_error"
-  | "signed_in";
-
-export type UserProfilePageProps = {
-  /** Page state (loading, signed out, …) */
-  state: UserProfilePageState;
-  /** All available skills */
-  allSkills: Skill[];
-  /** User profile for the signed-in user */
-  profile?: UserProfile;
-  /** What should be done when user clicks “Sign In” */
-  signIn: () => void;
-  /** What should be done when user clicks “Sign Out” */
-  signOut: () => void;
-  /** Called when selected skills change */
-  onUserSkillsChange: (skills: Skill[]) => void;
-};
-
-export const UserProfilePage: React.FC<UserProfilePageProps> = (props) => {
-  return (
-    <Layout crumbs={[{ label: "Profil uživatele" }]}>
-      <Section>
-        <SectionContent>
-          <PageContent {...props} />
-        </SectionContent>
-      </Section>
-    </Layout>
-  );
-};
-
-const PageContent: React.FC<UserProfilePageProps> = (props) => {
-  const { state, signIn } = props;
-  switch (state) {
-    case "signed_out":
-      return <SignedOutPage signIn={signIn} />;
-    case "loading":
-      return <LoadingPage />;
-    case "loading_error":
-      return <ErrorPage />;
-    case "signed_in":
-      return <SignedInPage {...props} />;
-  }
-};
-
-// Signed Out / Loading / Error
-
-const SignedOutPage: React.FC<{ signIn: () => void }> = ({ signIn }) => (
-  <MainContainer>
-    <Heading>Profil uživatele</Heading>
-    <Body>
-      Pro zobrazení této stránky se musíte{" "}
-      <a href="" onClick={signIn}>
-        přihlásit
-      </a>
-      .
-    </Body>
-  </MainContainer>
-);
-
-const LoadingPage = () => (
-  <MainContainer>
-    <Heading>Profil uživatele</Heading>
-    <Body>Načítám…</Body>
-  </MainContainer>
-);
-
-const ErrorPage = () => (
-  <MainContainer>
-    <Heading>Profil uživatele</Heading>
-    <Body>Při načítání profilu došlo k chybě. Zkuste obnovit stránku?</Body>
-  </MainContainer>
-);
-
-// User Profile
-
-const SignedInPage: React.FC<UserProfilePageProps> = (props) => {
-  const profile = props.profile!;
-  const { signOut, allSkills, onUserSkillsChange } = props;
-  return (
-    <MainContainer>
-      <Heading>{profile.name}</Heading>
-      <SkillBox
-        userSkillIds={profile.skills}
-        allSkills={allSkills}
-        onChange={onUserSkillsChange}
-      />
-      <Button
-        onClick={signOut}
-        style={{ marginTop: "40px", marginBottom: "40px" }}
-        inverted
-      >
-        Odhlásit
-      </Button>
-    </MainContainer>
-  );
-};
-
-type SkillBoxProps = {
+export type SkillBoxProps = {
   /** All available skills */
   allSkills: Skill[];
   /** IDs of the skills selected by user */
@@ -116,7 +13,7 @@ type SkillBoxProps = {
   onChange: (skills: Skill[]) => void;
 };
 
-const SkillBox: React.FC<SkillBoxProps> = ({
+export const SkillBox: React.FC<SkillBoxProps> = ({
   userSkillIds,
   allSkills,
   onChange,
@@ -211,16 +108,6 @@ const SkillPill: React.FC<SkillPillProps> = ({
     </AnimatedSpan>
   );
 };
-
-// Shared components
-
-const MainContainer: React.FC = (props) => (
-  <div style={{ minHeight: "400px" }}>{props.children}</div>
-);
-
-const Heading: React.FC = (props) => (
-  <Heading1 style={{ marginBottom: "20px" }}>{props.children}</Heading1>
-);
 
 const AnimatedSpan = styled.span`
   transition: transform 0.1s;
